requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==42.0.4
cycler==0.12.1
datefinder==0.7.3
Flask==3.0.2
fonttools==4.48.1
idna==3.6
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pycparser==2.21
pyOpenSSL==24.0.0
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
response==0.5.0
scipy==1.10.1
six==1.16.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
Werkzeug==3.0.1

